requirements.txt
flask
pillow
azure-storage-blob